R2 S4 V1

+++ src/main/java/sit/project/projectv1/controllers/AnnouncementController.java
@@ -4,12 +4,20 @@ import jakarta.validation.Valid;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import sit.project.projectv1.dtos.*;
-import sit.project.projectv1.entities.Announcement;
 import sit.project.projectv1.enums.Mode;
+import sit.project.projectv1.enums.Role;
+import sit.project.projectv1.exceptions.ItemNotFoundException;
+import sit.project.projectv1.models.Announcement;
+import sit.project.projectv1.models.User;
+import sit.project.projectv1.repositories.AnnouncementRepository;
 import sit.project.projectv1.services.AnnouncementService;
 import sit.project.projectv1.services.CategoryService;
+import sit.project.projectv1.services.UserService;
 import sit.project.projectv1.utils.ListMapper;
 
 import java.util.List;
@@ -24,6 +32,12 @@ public class AnnouncementController {
     private AnnouncementService announcementService;
 
     @Autowired
+    private AnnouncementRepository announcementRepository;
+
+    @Autowired
+    private UserService userService;
+
+    @Autowired
     private CategoryService categoryService;
 
     @Autowired
@@ -32,10 +46,14 @@ public class AnnouncementController {
     @Autowired
     private ListMapper listMapper;
 
+
     @GetMapping
     public List<AnnouncementDTO> getAnnouncementList(@RequestParam(defaultValue = "admin") Mode mode,
                                                      @RequestParam(defaultValue = "0") int category) {
-        List<Announcement> announcementList = announcementService.getAnnouncementList(mode, category);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUserFromToken(authentication);
+
+        List<Announcement> announcementList = announcementService.getAnnouncementList(mode, category, user);
         List<AnnouncementDTO> announcementDTOList = announcementList.stream()
                 .map(announcement -> modelMapper.map(announcement, AnnouncementDTO.class))
                 .collect(Collectors.toList());
@@ -47,13 +65,27 @@ public class AnnouncementController {
                                                            @RequestParam(defaultValue = "5") int size,
                                                            @RequestParam(defaultValue = "admin") Mode mode,
                                                            @RequestParam(defaultValue = "0") int category) {
-        Page<Announcement> announcementPage = announcementService.getAnnouncementPage(page, size, mode, category);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUserFromToken(authentication);
+
+        Page<Announcement> announcementPage = announcementService.getAnnouncementPage(page, size, mode, category, user);
         return listMapper.toPageDTO(announcementPage, AnnouncementDTO.class, modelMapper);
     }
 
     @GetMapping("/{announcementId}")
     public AnnouncementDetailDTO getAnnouncementById(@PathVariable Integer announcementId) {
-        return modelMapper.map(announcementService.getAnnouncementById(announcementId), AnnouncementDetailDTO.class);
+        Announcement storedAnnouncement = announcementRepository.findById(announcementId).orElseThrow(
+                () -> new ItemNotFoundException("This announcement does not exist!!!"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUserFromToken(authentication);
+
+        // Guest, admin, owner announcement can access
+        if (user == null || user.getRole() == Role.admin || user.getUsername().equals(storedAnnouncement.getAnnouncementOwner().getUsername())) {
+            return modelMapper.map(announcementService.getAnnouncementById(announcementId), AnnouncementDetailDTO.class);
+        }
+
+        throw new AccessDeniedException("Access denied!!!");
     }
 
     @GetMapping("/update/{announcementId}")
@@ -63,24 +95,50 @@ public class AnnouncementController {
 
     @DeleteMapping("/{announcementId}")
     public void deleteAnnouncementById(@PathVariable Integer announcementId) {
-        announcementService.deleteAnnouncementById(announcementId);
+        Announcement storedAnnouncement = announcementRepository.findById(announcementId).orElseThrow(
+                () -> new ItemNotFoundException("This announcement does not exist!!!"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUserFromToken(authentication);
+
+        if (user.getRole() == Role.admin || user.getUsername().equals(storedAnnouncement.getAnnouncementOwner().getUsername())) {
+            announcementService.deleteAnnouncementById(announcementId);
+        } else {
+            throw new AccessDeniedException("Access denied!!!");
+        }
     }
 
     @PostMapping
     public OutputAnnouncementDTO createAnnouncement(@Valid @RequestBody InputAnnouncementDTO announcementDTO) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUserFromToken(authentication);
+
         Announcement announcement = modelMapper.map(announcementDTO, Announcement.class);
         announcement.setId(null);
         announcement.setAnnouncementCategory(categoryService.getCategoryById(announcementDTO.getCategoryId()));
+        announcement.setAnnouncementOwner(user);
         announcementService.createAnnouncement(announcement);
         return modelMapper.map(announcement, OutputAnnouncementDTO.class);
     }
 
     @PutMapping("/{announcementId}")
     public OutputAnnouncementDTO updateAnnouncement(@PathVariable Integer announcementId, @Valid @RequestBody InputAnnouncementDTO announcementDTO) {
-        Announcement announcement = modelMapper.map(announcementDTO, Announcement.class);
-        announcement.setId(announcementId);
-        announcement.setAnnouncementCategory(categoryService.getCategoryById(announcementDTO.getCategoryId()));
-        announcementService.updateAnnouncement(announcementId, announcement);
-        return modelMapper.map(announcement, OutputAnnouncementDTO.class);
+        Announcement storedAnnouncement = announcementRepository.findById(announcementId).orElseThrow(
+                () -> new ItemNotFoundException("This announcement does not exist!!!"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUserFromToken(authentication);
+
+        if (user.getRole() == Role.admin || user.getUsername().equals(storedAnnouncement.getAnnouncementOwner().getUsername())) {
+            Announcement announcement = modelMapper.map(announcementDTO, Announcement.class);
+            announcement.setId(announcementId);
+            announcement.setAnnouncementCategory(categoryService.getCategoryById(announcementDTO.getCategoryId()));
+            announcement.setAnnouncementOwner(storedAnnouncement.getAnnouncementOwner());
+            announcementService.updateAnnouncement(announcementId, announcement);
+            return modelMapper.map(announcement, OutputAnnouncementDTO.class);
+        }
+
+        throw new AccessDeniedException("Access denied!!!");
     }
+
 }
